Split archive tab view into separate container component for use with methods
