+++ src/malcolmWidgets/method/methodViewer.component.js
@@ -16,6 +16,7 @@ import Typography from '@material-ui/core/Typography';
 
 // import ButtonAction from '../buttonAction/buttonAction.component';
 import WidgetTable from '../table/table.component';
+import MethodArchive from './methodArchive.container';
 
 import blockUtils from '../../malcolm/blockUtils';
 import { malcolmUpdateMethodInput } from '../../malcolm/actions/method.actions';
@@ -110,68 +111,14 @@ const MethodViewer = props => {
           );
         case 'widget:textinput':
         case 'widget:textupdate':
-        case 'widget:combobox': {
-          const timeStamps = props.methodArchive.timeStamp.toarray();
-          const values = props.methodArchive.value.toarray();
-          const dummyAttribute = {
-            raw: {
-              value: {
-                postTime: timeStamps.map(stamp => stamp.localRunTime),
-                returnTime: timeStamps.map(stamp => stamp.localReturnTime),
-                returnStatus: values.map(value => value.returnStatus),
-                alarm: props.methodArchive.alarmState.toarray(),
-                value: values.map(
-                  value => value.runParameters[props.selectedParam[1]]
-                ),
-              },
-              meta: {
-                elements: {
-                  alarm: {
-                    tags: ['info:alarm'],
-                    label: 'Method Alarm state',
-                  },
-                  postTime: {
-                    tags: ['widget:textupdate'],
-                    label: 'Time run',
-                  },
-                  returnTime: {
-                    tags: ['widget:textupdate'],
-                    label: 'Time results received',
-                  },
-                  returnStatus: {
-                    tags: ['widget:textupdate'],
-                    label: 'Method Return Status',
-                  },
-                  value: {
-                    tags: ['widget:textupdate'],
-                    label: 'Parameter value',
-                  },
-                },
-              },
-            },
-            calculated: {},
-          };
+        case 'widget:combobox':
+        case 'widget:checkbox':
+        case 'widget:led': {
           return (
-            <div className={props.classes.plainBackground}>
-              <div
-                className={props.classes.tableContainer}
-                style={{
-                  ...transitionWithPanelStyle,
-                  textAlign: 'left',
-                  display: 'initial',
-                }}
-              >
-                <WidgetTable
-                  attribute={dummyAttribute}
-                  hideInfo
-                  eventHandler={noOp}
-                  setFlag={noOp}
-                  addRow={noOp}
-                  infoClickHandler={noOp}
-                  rowClickHandler={noOp}
-                />
-              </div>
-            </div>
+            <MethodArchive
+              parentProps={props}
+              divStyle={transitionWithPanelStyle}
+            />
           );
         }
         default:
